requirements.txt
# =============================================================================
# Python Dependencies
# =============================================================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
starlette==0.35.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.26.0

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.1.0

# Data Processing
numpy==1.26.3
pandas==2.1.4

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2

# Frontend
streamlit==1.31.0
plotly==5.18.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
python-multipart==0.0.6
